# spec/swagger.yaml
...
swagger: '2.0'
schemes:
  - http
host: localhost:8813
basePath: /
info:
  description: |
    # A.L.E Double-Entry Accounting REST API

    ALE divides itself into "books", each of which store *journal entries* and their child *transactions*.
    The cardinal rule of double-entry accounting is that "everything must balance out to zero", and that rule is applied
    to every journal entry written to the book.
  version: '1.0.0'
  title: A.L.E.
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  x-logo:
    url: '/ale/logo.png'
...
produces:
  # This is a global default. You can OVERWRITE it in each specific operation.
  - application/json
consumes:
  # List of mime types the API endpoints consume
  # This is a global default. You can OVERWRITE it in each specific operation.
  - application/json
# Holds the relative paths to the individual endpoints. The path is appended to the basePath in order to construct the full URL.
paths:
...
  '/books/{bookId}/transactions':
    parameters:
    - name: bookId
      description: The book to extract entries from
      in: path
      type: integer
      required: true
    get:
      summary: List all transactions for given accounts
      operationId: getTransactions
      parameters:
      - name: accounts
        description: A comma-separated search term for accounts
        in: query
        type: string
        required: true

      - name: perPage
        in: query
        description: 'The number of results per page'
        required: false
        type: integer

      - name: page
        in: query
        description: 'The page number'
        required: false
        type: integer

      - name: newestFirst
        in: query
        description: 'If true (default - false), sort list with newest entries at the beginning'
        required: false
        type: boolean

      produces:
        - application/json
      responses: # list of responses
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Transaction'
...
definitions:
...
  Transaction:
    type: object
    required:
      - account
    properties:
      account:
        description: The account this transaction is reflected on
        type: string
      credit:
        description: The credit value of the transaction
        type: number
      debit:
        description: The debit value of the transaction
        type: number
      currency:
        description: The currency for this transaction
        type: string
      exchangeRate:
        description: The exchange rate to convert to the basis currency
        type: number
...